Module Bounds__Bound_unit

We view this bound as the single element `()`. In a sense, it is the worst over-approximation possible of anything. It maps to `0` when needed to convert to other numbers. This module is mainly useful for "meta abstract domain" that coordinates other abstract domains, but does not represent element or constraint. Design rational: The signature of `Abstract_domain` requires a `Bound`.

type t = unit

val name : string

val type_of : Core.Types.value_ty

val abstract_ty : Core.Types.var_abstract_ty

val concrete_ty : Core.Types.var_concrete_ty

val compare : 'a -> 'b -> int

val equal : 'a -> 'b -> bool

val leq : 'a -> 'b -> bool

val geq : 'a -> 'b -> bool

val lt : 'a -> 'b -> bool

val gt : 'a -> 'b -> bool

val neq : 'a -> 'b -> bool

val odd : 'a -> bool

val even : 'a -> bool

val min : 'a -> 'b -> 'a

val max : 'a -> 'b -> 'a

val sign : 'a -> int

val succ : 'a -> 'a

val prec : 'a -> 'a

val next_after : 'a -> 'b -> 'b

val of_int_up : 'a -> unit

val of_int_down : 'a -> unit

val of_float_up : 'a -> unit

val of_float_down : 'a -> unit

val of_rat_up : 'a -> unit

val of_rat_down : 'a -> unit

val of_string_up : 'a -> unit

val of_string_down : 'a -> unit

val to_string : 'a -> string

val output : Stdlib.out_channel -> 'a -> unit

val sprint : unit -> 'a -> string

val bprint : Stdlib.Buffer.t -> 'a -> unit

val pp_print : Stdlib.Format.formatter -> 'a -> unit

val to_float_up : 'a -> float

val to_float_down : 'a -> float

val to_rat : 'a -> Bounds.Bound_rat.t

val to_int_up : 'a -> int

val to_int_down : 'a -> int

type kind =

| FINITE

| MINF

| INF

| INVALID

val classify : 'a -> kind

val zero : unit

val one : unit

val two : unit

val minus_one : unit

val inf : unit

val minus_inf : unit

val neg : 'a -> unit

val abs : 'a -> unit

val add_up : 'a -> 'b -> unit

val sub_up : 'a -> 'b -> unit

val mul_up : 'a -> 'b -> unit

val div_up : 'a -> 'b -> unit

val add_down : 'a -> 'b -> unit

val sub_down : 'a -> 'b -> unit

val mul_down : 'a -> 'b -> unit

val div_down : 'a -> 'b -> unit

val bound_mul : 'a -> 'b -> 'c -> unit

val bound_div : 'a -> 'b -> 'c -> unit

val sqrt_up : 'a -> unit

val sqrt_down : 'a -> unit

val pow_up : 'a -> 'b -> unit

val pow_down : 'a -> 'b -> unit

val root_up : 'a -> 'b -> unit

val root_down : 'a -> 'b -> unit

val cos_up : 'a -> unit

val cos_down : 'a -> unit

val sin_up : 'a -> unit

val sin_down : 'a -> unit

val tan_up : 'a -> unit

val tan_down : 'a -> unit

val acos_up : 'a -> unit

val acos_down : 'a -> unit

val asin_up : 'a -> unit

val asin_down : 'a -> unit

val atan_up : 'a -> unit

val atan_down : 'a -> unit

val exp_up : 'a -> unit

val exp_down : 'a -> unit

val ln_up : 'a -> unit

val ln_down : 'a -> unit

val log_up : 'a -> unit

val log_down : 'a -> unit

val floor : 'a -> unit

val ceil : 'a -> unit