Module Box

module Box_split : sig ... end

module Box_interpretation : sig ... end

module Var_store : sig ... end

module type Box_sig = sig ... end

Box are an event-based abstract domain and must be encapsulated in an `Event_loop`. The `closure` operator is the identity function since it is decomposed in many tasks handled by `Event_loop`.

module type Box_functor = functor (B : Bounds.Bound_sig.S) -> Box_sig with module B = B and module Vardom.B = B

module Make : functor (B : Bounds.Bound_sig.S) -> functor (VARDOM : Vardom.Vardom_sig.Vardom_functor) -> functor (SPLIT : Box_split.Box_split_sig) -> Box_sig

module Box_base : functor (SPLIT : Box_split.Box_split_sig) -> Box_functor