Parameter Bisect.1-R

module Vardom : Vardom.Vardom_sig.S

module Store : Var_store.Var_store_sig with module V = Vardom

We depend on the store to represent the `var_id`. However, the store is contained in `Box` itself and not here. NOTE: We could parametrize Box_interpretation_sig with a `Store` if we have different kind of stores in the future.

include module type of sig ... end

type var_id = Store.key

type t

val empty : Typing.Ad_type.ad_uid -> t

val uid : t -> Typing.Ad_type.ad_uid

val extend : t -> (var_id * Typing.Tast.tvariable) -> t

val exists : t -> Lang.Ast.vname -> bool

val to_logic_var : t -> var_id -> Typing.Tast.tvariable

val to_abstract_var : t -> Lang.Ast.vname -> var_id * Typing.Tast.tvariable

val equantify : t -> Typing.Tast.tformula -> Typing.Tast.tqformula

val to_logic_var' : t -> var_id -> Lang.Ast.vname

val to_abstract_var' : t -> Lang.Ast.vname -> var_id

val local_vars : t -> Lang.Ast.vname -> var_id list

val to_abstract_var_wm : t -> Lang.Ast.vname -> var_id * Typing.Tast.tvariable

val interpret_gen : t -> string -> Typing.Tast.tformula -> (t -> Lang.Ast.bconstraint -> 'a list) -> t * 'a list

val interpret_gen' : t -> string -> Typing.Tast.tformula -> (t -> Typing.Tast.tformula -> t * 'a list) -> t * 'a list

val to_qformula_gen : t -> 'a list -> (t -> 'a -> Typing.Tast.tformula) -> Typing.Tast.tqformula

type var_dom = Store.cell

type rconstraint = var_id * var_dom

val exact_interpretation : bool

val interpret : t -> Domains.Interpretation.approx_kind -> Typing.Tast.tformula -> t * rconstraint list

val to_qformula : t -> rconstraint list -> Typing.Tast.tqformula