Module 1-R.Store

We depend on the store to represent the `var_id`. However, the store is contained in `Box` itself and not here. NOTE: We could parametrize Box_interpretation_sig with a `Store` if we have different kind of stores in the future.

type t

module V = Vardom

type cell = V.t

type key = int

val empty : t

val length : t -> int

val extend : ?⁠ty:Core.Types.var_ty -> t -> t * key * Core.Types.var_abstract_ty

val set : t -> key -> cell -> t

`set store k v` is a monotonic `set`. It merges `v` with `storek` using `V.meet`.

val get : t -> key -> cell

val lazy_copy : t -> int -> t list

val copy : t -> t

val iter : (key -> cell -> unit) -> t -> unit

val fold : ('a -> key -> cell -> 'a) -> 'a -> t -> 'a

val delta : t -> t * key list

This function consumes the registered delta in the store. The returned store has an empty list of delta.