Module Core__Tools

A small set of useful utilities

Printing stuff

val fail_fmt : ('a, Stdlib.Format.formatter, unit, 'b) Stdlib.format4 -> 'a

same as failwith but uses a format instead

val color_printf : Stdlib.Format.formatter -> string -> ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

terminal output with a color given in parameter restoring default color after use

val red_fprintf : Stdlib.Format.formatter -> ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

red terminal output

val cyan_fprintf : Stdlib.Format.formatter -> ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

blue terminal output

val green_fprintf : Stdlib.Format.formatter -> ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

green terminal output

val yellow_fprintf : Stdlib.Format.formatter -> ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

yellow terminal output

val matrix_print_indent : Stdlib.Format.formatter -> string array array -> unit

2D table print indentation

val pp_sep_newline : Stdlib.Format.formatter -> unit -> unit

Allocation of some of the most used printing utilities

val pp_sep_semicolon : Stdlib.Format.formatter -> unit -> unit

val pp_sep_comma : Stdlib.Format.formatter -> unit -> unit

val pp_sep_space : Stdlib.Format.formatter -> unit -> unit

val pp_sep_semicolon_space : Stdlib.Format.formatter -> unit -> unit

val pp_sep_comma_space : Stdlib.Format.formatter -> unit -> unit

val pp_print_float : Stdlib.Format.formatter -> float -> unit

Float printing utility

val pp_print_mpqf : Stdlib.Format.formatter -> Mpqf.t -> unit

Mpqf human understandable printing

val indent : string -> string

Start each new line with a two spaces indentation.

Conveniency functions

val range : int -> int -> int list

Create a list of integers between `i` and `j` included.

val is_some : 'a option -> bool

val unwrap : 'a option -> 'a

val fold_map : ('a -> 'b -> 'a * 'c) -> 'a -> 'b list -> 'a * 'c list

val filter_map : ('a -> 'b option) -> 'a list -> 'b list

val find_index : 'a -> 'a list -> int

val compare_list : ('a -> 'b -> bool) -> 'a list -> 'b list -> bool

val is_subset_list : 'a list -> 'a list -> bool

val is_set_equal : 'a list -> 'a list -> bool

val intersect : 'a list -> 'a list -> 'a list

val contain_string : string -> string -> (string -> int -> string) -> bool

val start_with : string -> string -> bool

True if `s` starts with `head`. `start_with` trims `s` and `head`.

val for_all_pairs : 'a list -> ('a -> 'a -> 'b list) -> 'b list

val indexed_list : 'a list -> (int * 'a) list

val for_all_distinct_pairs : 'a list -> ('a -> 'a -> 'b list) -> 'b list

val for_all_asymmetric_pairs : 'a list -> ('a -> 'a -> 'b list) -> 'b list

val fold_left_hd : ('a -> 'a -> 'a) -> 'a list -> 'a

Map instances

module VarMap : sig ... end

only one instanciation forall variable maps modules

val empty_parray : unit -> 'a Core.Parray.t

val extend_parray : 'a Core.Parray.t -> 'a -> 'a Core.Parray.t