Module Direct_product

This is a direct product combining N abstract domains A1...An. Whenever a variable or a constraint is typed with the UID of this product, it is added in ALL sub-domains. However, for functions operating on a single variable such as `to_abstract_var`, `project` or `embed`, only the first sub-domain is considered. To operate over a particular sub-domain, retrieve the variable ID with `local_vars`.

type gvar = Typing.Ad_type.ad_uid * int

Generic variables and constraints.

type gconstraint = Typing.Ad_type.ad_uid * int

module type Prod_combinator = sig ... end

type 'a owned_or_shared =

| Owned of 'a

| Shared of 'a

module Prod_atom : functor (A : Domains.Abstract_domain.Abstract_domain) -> sig ... end with type init_t = A.t Stdlib.ref owned_or_shared

module Prod_cons : functor (A : Domains.Abstract_domain.Abstract_domain) -> functor (B : Prod_combinator) -> Prod_combinator with type t = Prod_atom(A).t * B.t and type init_t = Prod_atom(A).init_t * B.init_t

module Direct_product : functor (P : Prod_combinator) -> sig ... end with module B = Bounds.Bound_rat