Module Domains.Abstract_domain

exception Conflict of int

Exception raised whenever a failure is encountered. In addition to `Bot.Bot_found`, `Conflict n` forces the backtrack in the search tree to the level `n`.

type task = Typing.Ad_type.ad_uid * int

type event = Typing.Ad_type.ad_uid * int

An event is usually a variable index of some sort.

type search_strategy =

| Simple

| VarView of Lang.Ast.vname list

| Sequence of (Typing.Ad_type.ad_uid * search_strategy) list

module type Abstract_domain = sig ... end

We require that abstract domains internalize their connection to the logical formula. In practice, it often consists in a pair `(A.t, A.I.t)` of the abstract element and its interpretation. The function `interpret` allows us to easily add constraints, the variables being added automatically if needed. See also `Interpretation.Interpretation_base`.