Module Interpretation.Interpretation_base

Parameters

V_ID : sig ... end

Signature

type var_id = V_ID.var_id

type t

val empty : Typing.Ad_type.ad_uid -> t

val uid : t -> Typing.Ad_type.ad_uid

val extend : t -> (var_id * Typing.Tast.tvariable) -> t

Add a mapping between a logical variable and its representation in the abstract domain.

val exists : t -> Lang.Ast.vname -> bool

`True` if the variable `v` exists in the abstract element. Note that it does not necessarily implies that a mapping exists.

val to_logic_var : t -> var_id -> Typing.Tast.tvariable

val to_abstract_var : t -> Lang.Ast.vname -> var_id * Typing.Tast.tvariable

val equantify : t -> Typing.Tast.tformula -> Typing.Tast.tqformula

Add existential quantifiers to the variables occuring in the formula.

val to_logic_var' : t -> var_id -> Lang.Ast.vname

Conveniency version of `to_logic_var` without the type of the variable.

val to_abstract_var' : t -> Lang.Ast.vname -> var_id

val local_vars : t -> Lang.Ast.vname -> var_id list

Same as `to_abstract_var'`.

val to_abstract_var_wm : t -> Lang.Ast.vname -> var_id * Typing.Tast.tvariable

Conveniency version of `to_abstrct_var` raising `Wrong_modelling` with a message indicating that the variable does not belong to the abstract element.