Module Domains__Interpretation

type approx_kind =

| Exact

| UnderApprox

| OverApprox

The kind of approximation when interpreting a logical formula into an abstract element. Note that the values `UnderApprox` or `OverApprox` interpret the constraint exactly if it is possible. Over-approximation: the abstract element might have more solutions than the concrete solutions of the formula. Under-approximation: the abstract element might have less solutions than the concrete solutions of the formula.

val neg_approx : approx_kind -> approx_kind

Over-approximation becomes under-approximation, and vice-versa.

val string_of_approx : approx_kind -> string

Past participle form of the verb representing the approximation (i.e. "under-approximated").

val no_variable_exn : string -> unit

Common error message for abstract domains that do not support variables (although their sub-domains might support variables).

module type Interpretation_sig = sig ... end

Every abstract domain may have a different variable and constraint representation according to their internal implementation. We ask every abstract domain to provide an interpretation module in order to connect the logic specification (`Ast.tqformula`) and the representation of the abstract domain. This module can also interpret a logic constraint into a more suited representation of the abstract domain. Rational: Why not directly pushing these functions into the definition of the abstract domain? We could parametrize the abstract domain by an interpretation although we do not do it yet. Another reason is that sometimes, we want to convert a formula into a abstract domain's constraint but without adding this formula in the abstract domain yet. This is for example the case with equivalence constraint "c1 <=> c2". We avoid the overhead of converting the logical to abstract constraint during the computation.

module Interpretation_base : functor (V_ID : sig ... end) -> sig ... end

val guarded_extend : 'a -> Typing.Ad_type.ad_uid -> string -> Typing.Tast.tvariable -> ('a -> Typing.Tast.tvariable -> 'a * 'b list) -> 'a * 'b list

guarded_extend a uid ad_name tv next is a helper function to extend domain with a variable. It checks that the formula's UID matches the one of the domain, and that the formula is not a tautology.

val guarded_interpret : 'a -> Typing.Ad_type.ad_uid -> string -> Typing.Tast.tformula -> ('a -> Typing.Tast.tformula -> 'a * 'b list) -> 'a * 'b list

guarded_interpret a uid ad_name tf next is a helper function to interpret a formula. It checks that the formula's UID matches the one of the domain, and that the formula is not a tautology.

val ground_interpret : 'a -> Typing.Ad_type.ad_uid -> string -> Typing.Tast.tformula -> ('a -> Lang.Ast.bconstraint -> 'b list) -> 'a * 'b list

ground_interpret a uid ad_name tf interpret_bconstraint traverses the formula and interpret every constraint. Raises Wrong_modelling if logic connectors other than conjunctions are present in the formula.

module Interpretation_ground : functor (V_ID : sig ... end) -> sig ... end

Helper functions to interpret and transform constraints in ground abstract domains.