Module Pengine.Pengine

type t

type event = int

For efficiency, we force the event and task ID to be integers.

type task_id = int

val empty : unit -> t

val extend_event : t -> t

Extend the reactor with a new event.

val extend_task : t -> t

Extend the reactor with a new task.

val react : t -> event list -> unit

Schedule all the tasks connected to at least one variable in the provided list.

val subscribe : t -> task_id -> event list -> t

Connect a task with a set of variables. If a variable is modified, the task is waken up.

val num_active_tasks : t -> int

The number of currently activated tasks.

val fixpoint : t -> ('a -> task_id -> 'a * bool * event list) -> 'a -> t * 'a

`fixpoint engine f acc` iterates over the scheduled tasks until no task are scheduled anymore. `f acc t` performs the task `t` over `acc` and returns the new `(acc, t_fixpoint, events)` where `acc` is the new accumulator, `t_fixpoint` is true if the task is at fixpoint, and `events` is a list of events to further react on.

val delta : t -> event list

This function consumes the registered delta in the engine. See `Var_store.delta`.