Module Fixpoint.Pengine2D

type t

type uid = int

type task = uid * int

type event = uid * int

val empty : unit -> t

val react : t -> event list -> unit

Schedule all the tasks connected to at least one event in the provided list. If the event does not exist in the engine, nothing happens. Rational: We suppose that no task is connected to that event.

val subscribe : t -> task -> event list -> t

Connect a task to a set of event. The task is waken up by `react` whenever one of the event occurs. The task and corresponding events are automatically allocated in the engine.

val num_active_tasks : t -> int

The number of currently activated tasks. An active task is a task that did not reach its solution space yet, and has more pruning to do.

val fixpoint : t -> ('a -> task -> 'a * bool * event list) -> 'a -> t * 'a * bool

`fixpoint engine f acc` iterates over the scheduled tasks until no task are scheduled anymore. `f acc t` performs the task `t` over `acc` and returns the new `(acc, is_entailed, events)` where `acc` is the new accumulator, `is_entailed` is true if the task reached its solutions space, and `events` is a list of events to further react on. It returns the new engine and accumulator, as well as a Boolean sets to `true` if at least an event was generated.