Module Transformer.Make

Parameters

A : Domains.Abstract_domain.Abstract_domain

Signature

type bab = {

kind : Lang.Ast.cmpop;

objective : A.I.var_id * Lang.Ast.vname;

best : (A.snapshot * Typing.Tast.tqformula) option;

}

`kind` is the type of BAB (minimization or maximization), see `minimize_bab` and `maximize_bab`. We keep the logical and abstract representation of the variable `objective` to optimize. `best` contains the best solution so far, or `None` if none was found yet.

type printer = {

print_node : string -> int -> A.t -> unit;

print_sol : A.t -> unit;

}

The arguments are as follows: `print_node status depth domain`. `print_node` is called in every node. `print_sol` is only called in solution nodes.

type transformer =

| Printer of printer

| BAB of bab

| Timeout of Mtime.span

| BoundSolutions of int

These are the transformers supported. They are compositional in the sense that we keep them in a list, and they are applied in turn when reaching a node.

val make_bab : Lang.Ast.cmpop -> (A.I.var_id * Lang.Ast.vname) -> transformer

See `type bab`.

val minimize_bab : (A.I.var_id * Lang.Ast.vname) -> transformer

val maximize_bab : (A.I.var_id * Lang.Ast.vname) -> transformer

type gs = {

transformers : transformer list;

stats : global_statistics;

domain : A.t;

}

Global state which stays identical to the whole computation.

type bs = {

snapshot : A.snapshot;

bt_stats : bt_statistics;

}

Backtrackable state which is automatically backtracked with the search tree.

type t = gs * bs

val unwrap_best : gs -> Bounds.Bound_rat.t option

Traverses the transformer stack and returns the best solution found so far, if any. Raises `Not_found` if no BAB transformer is found.

exception StopSearch of t

Any of the event functions `on_*` can raise this exception to stop the search.

exception Backjump of int * t

This exception prune the current subtree.

val wrap_exception : t -> (t -> t) -> t

Call `f t` and turn any `Bot_found` and `Conflict n` exception into a statefull `Backjump n t` exception. `n` is equal to `0` in case of `Bot_found`. Any function that can raise `Bot_found` or `Conflict` must be wrapped in this function. Rational: It prevents losing information on `t` obtained so far (such as the node counter).

val init : A.t -> transformer list -> t

val on_node : t -> t

Apply all the transformers modifying the current state when entering a node (before `closure`).

val on_fail : t -> t

Apply all the transformers concerned by a failed node.

val on_solution : t -> t

Apply all the transformers concerned by a solution node.

val on_unknown : t -> t

Apply all the transformers concerned by an unknown node.