Module Lang.Pretty_print

val print_unop : Stdlib.Format.formatter -> Ast.unop -> unit

val print_binop : Stdlib.Format.formatter -> Ast.binop -> unit

val print_cmpop : Stdlib.Format.formatter -> Ast.cmpop -> unit

val print_var : Stdlib.Format.formatter -> Ast.vname -> unit

val prior_level : (Ast.expr * Ast.expr) -> int

val must_parenthesis_right : Ast.expr -> Ast.expr -> bool

val print_expr : Stdlib.Format.formatter -> Ast.expr -> unit

This function prints an expression without redundant parenthesis and using continuation passing style to avoid a stack overflow error over very large constraints.

val print_formula : Stdlib.Format.formatter -> Ast.formula -> unit

val print_qformula : Stdlib.Format.formatter -> Ast.qformula -> unit

val print_constraint : Stdlib.Format.formatter -> Ast.bconstraint -> unit

val output_constraints : Ast.bconstraint list -> unit

val print_constraints : Stdlib.Format.formatter -> Ast.formula list -> unit

val string_of_constraint : Ast.bconstraint -> string

val string_of_formula : Ast.formula -> string