Module Lang__Ast

This file represents the AST of a constraint problem. This syntactic representation can be transformed by the `Representation_sig` of an abstract domain.

exception Wrong_modelling of string

This exception is raised when a constraint is passed to an abstract domain that cannot represent this constraint.

type vname = string

Variables are identified by a string.

type i = Bounds.Bound_rat.t

Constants are defined over rational because they can represent exactly integers or floating point numbers.

type unop =

| NEG

Unary arithmetic operators.

type binop =

| ADD

| SUB

| MUL

| DIV

| POW

Binary arithmetic operators.

type cmpop =

| EQ

| LEQ

| GEQ

| NEQ

| GT

| LT

Arithmetic comparison operators.

type expr =

| Funcall of string * expr list

| Unary of unop * expr

| Binary of expr * binop * expr

| Var of vname

| Cst of i * Core.Types.var_concrete_ty

type bconstraint = expr * cmpop * expr

A binary constraint.

type formula =

| FVar of vname

| Cmp of bconstraint

| Equiv of formula * formula

| Imply of formula * formula

| And of formula * formula

| Or of formula * formula

| Not of formula

Propositional logic formula.

type qformula =

| QFFormula of formula

| Exists of vname * Core.Types.var_ty * qformula

Formula with existential quantifier. It is mainly useful for declaring the concrete type of a variable's value.

type optimization_kind =

| Minimize of vname

| Maximize of vname

| Satisfy

type bab_qformula = {

qf : qformula;

optimise : optimization_kind;

}

Data structure to store a formula along with its variable to optimise.

val zero : expr

val one : expr

val two : expr

val truef : formula

Encoding of true and false as formula.

val falsef : formula

val has_variable : expr -> bool

Checks if an expression contains a variable.

val is_linear : expr -> bool

Checks if an expression is linear.

val is_cons_linear : formula -> bool

Checks if a constraint is linear.

val is_cst : expr -> bool

val type_dispatch : (module Bounds.Bound_sig.S) -> string -> Core.Types.var_ty -> (unit -> 'a) -> 'a

type_dispatch B from ty f. Given bound signature that is supported by an abstract domain, verify that the type is compatible with that signature, and call `f ()`. Raise `Wrong_modelling` in case of an incompatibility with a message including the name of the abstract domain `from`.