Module Cumulative.MakeTimeRD

Time-resource decomposition of cumulative. For this decomposition, task variables are required to be discrete as we need to discretize over time. `C` is the bound of resources / capacities.

Parameters

C : Bounds.Bound_sig.S

Signature

include module type of sig ... end

module D : sig ... end

type param_or_var = Lang_decompositions__Scheduling.Make(Bounds.Bound_int).param_or_var =

| Param of D.t

| Variable of string

type task = Lang_decompositions__Scheduling.Make(Bounds.Bound_int).task = {

start : string;

duration : param_or_var;

}

val non_overlap : task -> task -> Lang.Ast.formula

val overlap_before : task -> task -> Lang.Ast.formula

val disjunctive : task list -> Lang.Ast.formula

val at_instant : task -> param_or_var -> Lang.Ast.formula

val precedence : string -> string -> param_or_var -> Lang.Ast.formula

module C : sig ... end

type rtask = Generic(Bounds.Bound_int)(C).rtask = {

task : task;

id : int;

resources_usage : C.t;

}

type name_factory = rtask -> int -> string

Given a task `t` and a time instant `i`, get a fresh or existing name of the Boolean variable representing that `t` is running at the instant `i`.

val default_name_factory : name_factory

We have `task_1_runs_at_2 = 1` if the task `1` runs at the instant `2`.

val shared_constraints : rtask list -> D.t -> name_factory -> Lang.Ast.qformula

Similar to `MakeTaskRD.shared_constraints`. Here the decomposition is: `forall(i,t) task_<t.id>_runs_at_<i> <=> at_instant(t.task, i).

val cumulative : rtask list -> D.t -> C.t -> name_factory -> Lang.Ast.formula

Time-resource decomposition of cumulative. Similar to `MakeTaskRD.cumulative`. `horizon` is an upper bound on the latest possible time. The decomposition is `forall(i in 0..horizon) capacity >= sum(t) task_<t.id>_runs_at_<i> * t.resource_usage