Module Lang_decompositions__Cumulative

Decompositions of the cumulative global constraint.

module type Cumulative_decomposition = sig ... end

module Generic : functor (T : Bounds.Bound_sig.S) -> functor (C : Bounds.Bound_sig.S) -> sig ... end with module C = C

module MakeTaskRD : functor (T : Bounds.Bound_sig.S) -> functor (C : Bounds.Bound_sig.S) -> sig ... end

`T` is the bound of task and `C` is the bound of resources / capacities.

module MakeTimeRD : functor (C : Bounds.Bound_sig.S) -> sig ... end

Time-resource decomposition of cumulative. For this decomposition, task variables are required to be discrete as we need to discretize over time. `C` is the bound of resources / capacities.