Module Lang_decompositions__Cumulative.MakeTaskRD

`T` is the bound of task and `C` is the bound of resources / capacities.

Parameters

T : Bounds.Bound_sig.S

C : Bounds.Bound_sig.S

Signature

include module type of sig ... end

module D : sig ... end

type param_or_var = Lang_decompositions__Scheduling.Make(T).param_or_var =

| Param of D.t

| Variable of string

type task = Lang_decompositions__Scheduling.Make(T).task = {

start : string;

duration : param_or_var;

}

val non_overlap : task -> task -> Lang.Ast.formula

val overlap_before : task -> task -> Lang.Ast.formula

val disjunctive : task list -> Lang.Ast.formula

val at_instant : task -> param_or_var -> Lang.Ast.formula

val precedence : string -> string -> param_or_var -> Lang.Ast.formula

module C : sig ... end

type rtask = Generic(T)(C).rtask = {

task : task;

id : int;

resources_usage : C.t;

}

type name_factory = rtask -> rtask -> string

Given two tasks `t1` and `t2`, get a fresh or existing name of the Boolean variable representing that `t1` is running when `t2` starts.

val default_name_factory : name_factory

We have `task_1_runs_when_2_starts = 1` if the task `1` starts when the task `2` is running.

val shared_constraints : rtask list -> int -> name_factory -> Lang.Ast.qformula

Create constraints that are shared among the task-RD cumulative decomposition. `MakeTaskRD.cumulative` suppose these constraints are available. Any `resources_usage` in tasks can be used since it is not used to generate these constraints. `forall(t1 <> t2) task_<t2.id>_runs_when_<t1.id>_starts <=> overlap_before(t1,t2).

val cumulative : rtask list -> D.t -> C.t -> name_factory -> Lang.Ast.formula

Task-resource decomposition of cumulative. You must also retrieve the constraints of `shared_constraints`. Tasks not using any resource are automatically discarded. `name_factory` helps to share new Boolean variables created by the decomposition across all cumulatives. The decomposition is `forall(t1), capacity - t1.resource_usage >= sum(t2 where t1 <> t2) (task_<t2.id>_runs_when_<t1.id>_starts * t2.resource_usage)`