Module type Lang_decompositions__Cumulative.Cumulative_decomposition

include Lang_decompositions.Scheduling.S

module D : Bounds.Bound_sig.S

type param_or_var =

| Param of D.t

| Variable of string

type task = {

start : string;
Name of the variable representing the starting date.

duration : param_or_var;
Duration of a task, it can either be a parameter or a variable.

}

val non_overlap : task -> task -> Lang.Ast.formula

Ensure two tasks do not overlap. The resulting decomposition is `s1 - s2 <= -d1 \/ s2 - s1 <= -d2`.

val overlap_before : task -> task -> Lang.Ast.formula

Ensure that `t1` overlaps with `t2` and that `t1` starts before `t2`. The resulting decomposition is `s1 <= s2 /\ s2 - s1 < d1

val disjunctive : task list -> Lang.Ast.formula

Ensure that `tasks` are never scheduled at the same time (they do not overlap).

val at_instant : task -> param_or_var -> Lang.Ast.formula

at_instant t i Ensure that a task `t` is running at the instant `i`.

val precedence : string -> string -> param_or_var -> Lang.Ast.formula

Given two tasks name `s1,s2`, and a duration `d`, ensure that `s2` starts at least `d` units after the start of `s1`. The resulting constraint is `s1 + d1 <= s2` that we rewrite to `s1 - s2 <= -d1`.

module C : Bounds.Bound_sig.S

type rtask = {

task : task;

id : int;
The ID of a task is useful to create name of new variables involving this task. It should uniquely identify the task.

resources_usage : C.t;

}

A task with resource consumption.

type name_factory

val default_name_factory : name_factory

val shared_constraints : rtask list -> D.t -> name_factory -> Lang.Ast.qformula

val cumulative : rtask list -> D.t -> C.t -> name_factory -> Lang.Ast.formula