Module Logic_completion

`Logic_completion` equips an abstract domain with support for logical formula. The main operation required on the underlying abstract domain is `entailment`. It generalizes to any propositional logical formula the results presented in the paper: "Combining Constraint Languages via Abstract Interpretation" (Talbot and al., 2019). It is also related to `Disjunctive completion` in abstract interpretation.

module type Logic_completion_interpretation_sig = sig ... end

Variables cannot be represented in this abstract domain. They must be interpreted through the corresponding sub-domain or in another reduced product.

module Logic_completion : functor (A : Domains.Abstract_domain.Abstract_domain) -> sig ... end with module B = Bounds.Bound_unit and module I.A = A