Module OctagonF.I

The representation of the variables and constraints inside the abstract element. It allows users to turn a logic specification into an abstract element.

type t

type var_id

Variable ID as represented in the abstract domain.

type rconstraint

Constraint representation in the abstract domain.

val exact_interpretation : bool

`true` if `interpret` always exactly interpret constraint (or fail), false if under- or over-approximations are possible.

val empty : Typing.Ad_type.ad_uid -> t

An empty interpretation.

val to_logic_var : t -> var_id -> Typing.Tast.tvariable

Conversion between abstract variables and their logical representations.

val to_abstract_var : t -> Lang.Ast.vname -> var_id * Typing.Tast.tvariable

Conversion between a logical variables and its abstract representation. Raises `Not_found` if the variable is not in `repr`.

val local_vars : t -> Lang.Ast.vname -> var_id list

local_vars repr v gives the set of local variables representing `v` in the abstract element `repr`. Same behavior than `to_abstract_var` if the variable has none or a single local variable. In case of multiple variables, these variables must be duplicate, in the sense that they represent the same initial variable (possibly with different level of approximation). For instance, a variable `v` typed in a direct product "A1 X A2" has two versions: one in A1, one in A2.

val interpret : t -> Domains.Interpretation.approx_kind -> Typing.Tast.tformula -> t * rconstraint list

Interpret a logic formula into a list of abstract constraints. It approximates the representation of the formula if needed according to `approx`. The list of constraints might be empty if the formula is detected to be a tautology. Raise `Wrong_modelling` if: 1. the formula cannot be approximated according to `approx` in the abstract domain. 2. a variable of the formula does not belong to the interpretation.

val to_qformula : t -> rconstraint list -> Typing.Tast.tqformula

Give a logical representation of an abstract element. This function is the reverse of `interpret`. Note that we do not need to approximate the result as a formula should always be able to represent exactly an element. The free variables in the formula obtained should be considered existentially quantified.