Module type Octagon_split.ValueDistributor

This module aggregates `Value_order` and `Distributor`. It is often safer to have both in a single function because of incompatibilities. For example, the lower bound value must be bisected with `x <= lb \/ x > lb` instead of `x < lb \/ x >= lb`.

Parameters

DBM : Dbm.DBM_sig

Signature

module DBM = DBM

val distribute : DBM.t -> Dbm.dbm_interval -> DBM.bound Dbm.dbm_constraint list