Module Octagon__.Dbm

type dbm_var = {

l : int;

c : int;

}

A variable is represented by its position in the DBM (line, column).

type 'b dbm_constraint = {

v : dbm_var;

d : 'b;

}

Constraint of the form `±x - ±y <= d`.

type dbm_interval = {

lb : dbm_var;

ub : dbm_var;

}

An interval view of a variable in the DBM with its lower and upper bounds.

val make_var : int -> int -> dbm_var

Create a variable position in the DBM by taking into account the coherence.

val make_canonical_var : int -> dbm_var

Given a dimension `k`, it gives a canonical representation of this variable in the DBM (which is arbitrarily the lower bound of the interval).

val inv : dbm_var -> dbm_var

If `v` is the position of a lower bound, it returns the position of its associated upper bound, and conversly. For example given `x - y`, it returns the variable representing `-x + y` which is represented by the element in its diagonal on the same plane.

val is_lower_bound : dbm_var -> bool

val is_upper_bound : dbm_var -> bool

val as_interval : dbm_var -> dbm_interval

val is_rotated : dbm_var -> bool

`true` if the variable is in a rotated plane.

val event_of_var : dbm_var -> int list

Retrieve the event index of a variable.

module type Fold_interval_sig = sig ... end

module Fold_intervals : Fold_interval_sig

module Fold_intervals_canonical : Fold_interval_sig

module Fold_intervals_rotated : Fold_interval_sig

module type DBM_sig = sig ... end

module Make : functor (B : Bounds.Bound_sig.S) -> DBM_sig

module MakeCopy : functor (B : Bounds.Bound_sig.S) -> DBM_sig

module MakeTrailing : functor (B : Bounds.Bound_sig.S) -> DBM_sig