Module Octagon__.Octagon_split

module type Variable_order = functor (DBM : Octagon.Dbm.DBM_sig) -> sig ... end with module DBM = DBM

module type Value_order = functor (DBM : Octagon.Dbm.DBM_sig) -> sig ... end with module DBM = DBM

module type Distributor = functor (DBM : Octagon.Dbm.DBM_sig) -> sig ... end with module DBM = DBM

module type ValueDistributor = functor (DBM : Octagon.Dbm.DBM_sig) -> sig ... end with module DBM = DBM

This module aggregates `Value_order` and `Distributor`. It is often safer to have both in a single function because of incompatibilities. For example, the lower bound value must be bisected with `x <= lb \/ x > lb` instead of `x < lb \/ x >= lb`.

module type Octagon_split_sig = functor (DBM : Octagon.Dbm.DBM_sig) -> sig ... end with module DBM = DBM

module Middle : Value_order

module Input_order : functor (Fold_interval : Octagon.Dbm.Fold_interval_sig) -> Variable_order

module First_fail : functor (Fold_interval : Octagon.Dbm.Fold_interval_sig) -> Variable_order

module Anti_first_fail : functor (Fold_interval : Octagon.Dbm.Fold_interval_sig) -> Variable_order

module Min_LB_AFF_IO : functor (Fold_interval : Octagon.Dbm.Fold_interval_sig) -> Variable_order

module Min_LB : functor (Fold_interval : Octagon.Dbm.Fold_interval_sig) -> Variable_order

module Min_UB : functor (Fold_interval : Octagon.Dbm.Fold_interval_sig) -> Variable_order

module Max_LB : functor (Fold_interval : Octagon.Dbm.Fold_interval_sig) -> Variable_order

module Max_UB : functor (Fold_interval : Octagon.Dbm.Fold_interval_sig) -> Variable_order

module Max_min : Variable_order

module Min_max : Variable_order

module Bisect_middle : ValueDistributor

module Assign_LB : ValueDistributor

module Assign_UB : ValueDistributor

module Right_to_left : functor (DISTRIBUTOR : ValueDistributor) -> ValueDistributor

module Make : functor (VARIABLE : Variable_order) -> functor (VALUE_DISTRIBUTOR : ValueDistributor) -> Octagon_split_sig

module First_fail_bisect : Octagon_split_sig

module Min_max_LB : Octagon_split_sig

module Max_min_LB : Octagon_split_sig

module Max_min_UB : Octagon_split_sig

module Max_min_Bisect : Octagon_split_sig

module Min_max_Bisect : Octagon_split_sig

module Max_min_Bisect_reverse : Octagon_split_sig

module First_fail_LB_canonical : Octagon_split_sig

module Anti_first_fail_LB_canonical : Octagon_split_sig

module Anti_first_fail_UB_canonical : Octagon_split_sig

module Anti_first_fail_LB : Octagon_split_sig

module Anti_first_fail_Bisect : Octagon_split_sig

module Anti_first_fail_UB : Octagon_split_sig

module MSLF_simple : Octagon_split_sig

module MSLF_simple_rotated : Octagon_split_sig

module MSLF : Octagon_split_sig

module MSLF_all : Octagon_split_sig

module MSLF_rotated : Octagon_split_sig

module MSLF_UB : Octagon_split_sig

module MSLF_UB_all : Octagon_split_sig