Parameter Min_max_LB.1-DBM

module B : Bounds.Bound_sig.S

type bound = B.t

type t

val init : int -> t

Initialize a DBM of dimension `n`.

val empty : t

Initialize an empty DBM.

val extend : ?⁠ty:Core.Types.var_ty -> t -> t * Octagon.Dbm.dbm_var * Core.Types.var_abstract_ty

Increase the dimension of the DBM by one. We return the interval of the canonical interval in this dimension.

val get : t -> Octagon.Dbm.dbm_var -> bound

Low level access to a cell of the DBM where `get m l c` returns DBMlc. Precondition: `c/2 <= l/2` (always ensured if `dbm_var` is built with `make_var`).

val set : t -> bound Octagon.Dbm.dbm_constraint -> t

Monotonic write: we update the cell at (l,c) only if the value passed as argument is smaller than the one in the DBM. Precondition on the variable: same as `get`.

val project : t -> Octagon.Dbm.dbm_interval -> bound * bound

Returns the interval value of a pair of DBM variables. Precisely, it returns (-dbmlb, dbmub). See also `Interval_view_dbm` and `Octagon.project` to recover an "interval" interpretation of these bounds.

val lazy_copy : t -> int -> t list

Return `n` copies of the current DBM with the assumption that this one will not be used anymore. See `Abstract_domain.lazy_copy`.

val copy : t -> t

val dimension : t -> int

The dimension of the DBM is its number of variables.

val to_list : t -> bound list

Low-level representation of the DBM as a list.

val print : Stdlib.Format.formatter -> t -> unit

See `DBM.print`.

val delta : t -> t * int list

The events returned correspond to the position of the variables in the DBM.

val has_changed : t -> bool

True if some deltas were created.