Module Octagon__Octagon_interpretation

This module provides functions to detect and rewrite arbitrary constraints into an equivalent and relaxed octagonal constraints, if possible.

module type Octagon_interpretation_sig = sig ... end

module Octagon_interpretation : functor (B : Bounds.Bound_sig.S) -> Octagon_interpretation_sig with module B = B