Module Propagator_completion

module PC_interpretation = Propagator_completion__.Pc_interpretation

module Hc4 : sig ... end

module type Propagator_completion_sig = sig ... end

Propagator_completion is an event-based abstract domain and must be encapsulated in an `Event_loop`. The `closure` operator is the identity function since it is decomposed in many tasks handled by `Event_loop`.

module Propagator_completion : functor (V : Vardom.Vardom_sig.S) -> functor (A : Domains.Abstract_domain.Abstract_domain) -> Propagator_completion_sig with module V = V and module B = V.B and module I.A = A