Parameter Make.1-I

module V : Vardom.Vardom_sig.S

module A : Domains.Abstract_domain.Abstract_domain

type t = {

a : A.t Stdlib.ref;

uid : Typing.Ad_type.ad_uid;

}

type var_id = A.I.var_id list

type rexpr = {

node : node;

mutable value : V.t;

}

We annotate each node of this expression with its interval evaluation. This is useful for the HC4 algorithm. The `value` field is get_vars_set_formulanever backtracked, it is only useful to propagate inside on node of the search tree.

and node =

| BFuncall of string * rexpr list

| BUnary of Lang.Ast.unop * rexpr

| BBinary of rexpr * Lang.Ast.binop * rexpr

| BVar of var_id * Typing.Tast.tvariable

| BCst of V.t * Core.Types.var_abstract_ty

type rconstraint = rexpr * Lang.Ast.cmpop * rexpr

val name : string

val exact_interpretation : bool

val empty : Typing.Ad_type.ad_uid -> t

val to_logic_var : t -> var_id -> Typing.Tast.tvariable

val to_abstract_var : t -> Lang.Ast.vname -> var_id * Typing.Tast.tvariable

val local_vars : t -> Lang.Ast.vname -> var_id list

val interpret : t -> Domains.Interpretation.approx_kind -> Typing.Tast.tformula -> t * rconstraint list

Over-approximation is always supported. No special effort is currently made for supporting under-approximations, it is considered equivalent to Exact representation. Exact representation is supported if the abstract types of the variables of `f` are all different from floating point numbers.

val to_qformula : t -> rconstraint list -> Typing.Tast.tqformula

val make_expr : node -> rexpr

Create an expression from a node. The vardom ref is initialized to TOP.