Module Sat

This module encapsulates a CDCL SAT solver based on MiniSat, and implemented by Albin Coquereau in OCaml in a library called Minisatml. WARNING: only one instance of this module can be used at once, this is because Minisatml relies on global variables.

module Sat : Domains.Abstract_domain.Abstract_domain with module B = Bounds.Bound_int and module I = Sat__.Sat_interpretation.Sat_interpretation