Module Typing__Aast

Annotated AST (AAST). This AST is used internally by the type inference engine (`Infer.ml`).

type 'a aformula = 'a * 'a aformula_

Generic annotated version of a formula. It is useful to have this generic structure for UID types, and in Infer where we need to store local information along the formula.

and 'a aformula_ =

| AFVar of Lang.Ast.vname

| ACmp of Lang.Ast.bconstraint

| AEquiv of 'a aformula * 'a aformula

| AImply of 'a aformula * 'a aformula

| AAnd of 'a aformula * 'a aformula

| AOr of 'a aformula * 'a aformula

| ANot of 'a aformula

type 'a aqformula =

| AQFFormula of 'a aformula

| AExists of Lang.Ast.vname * Core.Types.var_ty * 'a * 'a aqformula

val aformula_to_formula : 'a aformula -> Lang.Ast.formula

Remove the types from a formula. It is particularly useful to use the functions available in Rewriting and Pretty_print modules.

val aqformula_to_qformula : 'a aqformula -> Lang.Ast.qformula

val string_of_aformula : 'a aformula -> string

val map_tqf : ('a aformula -> 'a aformula) -> 'a aqformula -> 'a aqformula

val map_annot_aformula : 'a aformula -> ('a aformula -> 'b) -> 'b aformula

val map_annot_aqformula : 'a aqformula -> ('a -> 'b) -> 'b aqformula