Module Typing__Ad_type

Symbolic representation of the abstract domain by their "types". For now, only the most interesting combination are represented, for instance I left out BDD because it is not complete enough to be useful. These types should be extended whenever a new domain or combination is added.

See also the Typing module.

type ad_uid = int

type vardom_ty =

| Interval of Core.Types.value_ty

| Interval_oc of Core.Types.value_ty

| Interval_mix

type ad_ty_ =

| Box of vardom_ty

| Octagon of Core.Types.value_ty

| SAT

| Logic_completion of ad_ty

| Propagator_completion of ad_ty

| Cascade_product of ad_ty * ad_ty

| Direct_product of ad_ty list

and ad_ty = ad_uid * ad_ty_

val string_of_vardom_ty : vardom_ty -> string

val string_of_adty : ad_ty -> string

val is_more_specialized_value : Core.Types.value_ty -> Core.Types.value_ty -> Core.Kleene.t

Rational is more specialized than float (because more precise). Integers and rational/float are unordered.

val is_more_specialized_vardom : vardom_ty -> vardom_ty -> Core.Kleene.t

val subtype : ad_ty -> ad_ty -> bool

subtype a b is true if a is included in b. Example: Box is included in PC(Box X Oct).

val is_more_specialized : ad_ty -> ad_ty -> Core.Kleene.t

Give an order over the abstract domains. An abstract domain A1 is more specialized than A2 if: (1) The constraint language of A1 is included in A2, and (2) A1 can infer more information than A2 (regardless of efficiency). We have: SAT > Octagon > Box. Inductive call on Logic_transformer and Direct_product.

module UID2Adty : Stdlib.Map.S with type UID2Adty.key = ad_uid

val ad_name : ad_ty UID2Adty.t -> ad_uid -> string

Find and format the name of an abstract domain. It is used in the header of the error messages.

val string_of_type : ad_ty UID2Adty.t -> ad_uid -> string

val string_of_adty_env : ad_ty UID2Adty.t -> string

val build_adenv : ad_ty -> ad_ty UID2Adty.t

val select_mgad : ad_ty list -> ad_ty option

Filter the abstract domains in the list which subsume all the others. The list is empty if none exists.

val is_mgad : ad_ty -> ad_uid list -> bool

True if `adty` supports all the domains `uids`