Module Typing__Infer

This module attempt to give a type to each component of a logical formula. In our context, a type is a list of UIDs representing the abstract domains in which the part of a formula can be typed.

NOTE: Only the function infer_type should be useful for users of this module, but we also give the functions of the inference engine for documentation purposes.

type inferred_type =

| CannotType of string

| Typed of Typing.Ad_type.ad_uid list

The type inferred from a formula. In case the user already provided some UIDs, we flag these with a Boolean.

val merge_ity : inferred_type -> inferred_type -> inferred_type

type iformula = inferred_type Typing.Aast.aformula

We equip a formula with the UIDs of abstract elements in which it can be interpreted. The UID `0` is by default the largest abstract type avalaible (usually a product among the available abstract domains).

type iqformula = inferred_type Typing.Aast.aqformula

val uids_of' : inferred_type -> Typing.Ad_type.ad_uid list

val uids_of : iformula -> Typing.Ad_type.ad_uid list

val is_uid_in : Typing.Ad_type.ad_uid -> iformula -> bool

val is_uid_in2 : Typing.Ad_type.ad_uid -> iformula -> iformula -> bool

val formula_to_iformula : Lang.Ast.formula -> iformula

See qformula_to_iqformula.

val qformula_to_iqformula : Lang.Ast.qformula -> iqformula

Convert an untyped formula to a formula with a empty type (Typed ) that will be populated.

module Inference : sig ... end

Module containing all the necessary information to type a formula with a given abstract domain type.

val make_tqformula : Typing.Ad_type.ad_uid Typing.Aast.aqformula -> Typing.Tast.tqformula

Final step of the typing process where the formula is cleaned with the UID information.

val infer_type : Typing.Ad_type.ad_ty -> Lang.Ast.qformula -> Typing.Tast.tqformula

Given an abstract domain type and an untyped formula, infer the type of each component of the formula matching the given abstract domain. Raises `Wrong_modelling msg` if the formula is not typable for this abstract domain's type; `msg` should explain why the formula is not typable. Note: `ad_ty` encompasses domain transformers including product such as "t1 X t2", thus part of a formula might be typed with `t1` or `t2`.