Module Typing__Infer.Inference

Module containing all the necessary information to type a formula with a given abstract domain type.

module Var2UID : Stdlib.Map.S with type Var2UID.key = Lang.Ast.vname

type var_env = Typing.Ad_type.ad_uid list Var2UID.t

type t

val init : Typing.Ad_type.ad_ty -> bool -> t

Boolean is the `trace` which indicates if the inference should create the error message or not (it might take longer if yes).

val gen_err : t -> Typing.Ad_type.ad_uid -> (unit -> string) -> inferred_type

Create `CannotType` with the given message if the trace is activated, and an empty message otherwise.

val create_typing_error : string -> iformula -> string

Format the typing error of a formula. We try to give the most precise error (an error on a small sub-formula).

val variable_not_in_dom_err : t -> Typing.Ad_type.ad_uid -> Lang.Ast.vname -> inferred_type

val variable_not_in_subdom_err : t -> Typing.Ad_type.ad_uid -> Typing.Ad_type.ad_uid -> Lang.Ast.vname -> inferred_type

val not_a_box_constraint_err : t -> Typing.Ad_type.ad_uid -> inferred_type

val not_an_octagonal_constraint_err : t -> Typing.Ad_type.ad_uid -> inferred_type

val ground_dom_does_not_handle_logic_connector_err : t -> Typing.Ad_type.ad_uid -> inferred_type

val no_domain_support_this_variable_err : t -> Typing.Ad_type.ad_uid -> Lang.Ast.vname -> Core.Types.var_ty -> inferred_type

val sat_does_not_support_term_err : t -> Typing.Ad_type.ad_uid -> inferred_type

val direct_product_no_subdomain_err : t -> Typing.Ad_type.ad_uid -> string -> inferred_type

val logic_completion_subdomain_failed_on_term_err : t -> Typing.Ad_type.ad_uid -> inferred_type

val interval_can_represent_var : Typing.Ad_type.vardom_ty -> Core.Types.var_ty -> bool

val compatible_ty : Core.Types.var_ty -> Core.Types.value_ty -> bool

val infer_var : Core.Types.var_ty -> Typing.Ad_type.ad_ty -> Typing.Ad_type.ad_uid list

val infer_vars_ty : t -> iqformula -> t * iqformula

For each variable, infer its UIDs.

val belong : t -> Typing.Ad_type.ad_uid -> Lang.Ast.vname -> bool

val bool_var_infer : t -> Lang.Ast.vname -> Typing.Ad_type.ad_uid -> inferred_type

val ground_dom_infer : t -> Typing.Ad_type.ad_uid -> (Lang.Ast.bconstraint -> inferred_type) -> iformula -> iformula

Infer type for ground abstract domain, term_infer gives a type to term. It is a helper to implement inference for box, octagon and other ground abstract domains. It walks through conjunction `TAnd`, `TFVar` and call term_infer on `TCmp`.

val fully_defined_over : t -> Typing.Ad_type.ad_uid -> Lang.Ast.bconstraint -> Lang.Ast.vname option

`None` if the variables of `c` are all treatable in the element `uid`, otherwise `Some v` where `v` is a variable not treatable.

val box_infer : t -> Typing.Ad_type.ad_uid -> Typing.Ad_type.vardom_ty -> iformula -> iformula

Infer the type of all the box constraints.

val octagon_infer : t -> Typing.Ad_type.ad_uid -> iformula -> iformula

Infer the type of all the octagonal constraints.

val generic_formula_infer : t -> Typing.Ad_type.ad_uid -> iformula -> (Lang.Ast.vname -> iformula -> inferred_type) -> (Lang.Ast.bconstraint -> iformula -> inferred_type) -> iformula

A generic inference scheme for logical formula (useful for SAT and Logic completion). literal and term are functions called on TFVar and TCmp respectively.

val sat_infer : t -> Typing.Ad_type.ad_uid -> iformula -> iformula

Infer the type of all SAT subformulas.

val direct_product_infer : t -> Typing.Ad_type.ad_uid -> iformula -> Typing.Ad_type.ad_ty list -> iformula

Infer the type of all subformulas compatible with this direct product.

val logic_completion_infer : t -> Typing.Ad_type.ad_uid -> iformula -> Typing.Ad_type.ad_ty -> iformula

Infer the type of all subformulas compatible with the logic completion of the given domain's type.

val propagator_completion_infer : t -> Typing.Ad_type.ad_uid -> iformula -> Typing.Ad_type.ad_ty -> iformula

Infer the type of all constraints compatible with the propagator completion of the given domain's type.

val infer_constraints_ty : t -> iformula -> Typing.Ad_type.ad_ty -> iformula

General inference of a formula with regards to the available abstract domain type.

val infer_constraints_ty_or_fail : t -> iqformula -> iqformula

Type a formula or raise `Wrong_modelling` if it cannot type it.

module VarConsCounter : Stdlib.Map.S with type VarConsCounter.key = Lang.Ast.vname * Typing.Ad_type.ad_uid

val build_var_cons_map : 'a Typing.Aast.aformula -> ('a -> int list) -> (int * int) VarConsCounter.t

Create a structure storing the number of unary and nary constraints for each variable and domain uid.

val restrict_unary_var_dom : t -> Typing.Ad_type.ad_uid list -> Typing.Ad_type.ad_uid list

This part is quite specialized, we need more research to infer better general rules. Basically, we do not want to add an unary variable in an octagon if a box (or SAT) is available, because it is less costly.

val restrict_variable_ty : t -> iqformula -> t * iqformula

Instantiate a variable with a single type.

val instantiate_formula_ty : t -> iformula -> Typing.Ad_type.ad_uid Typing.Aast.aformula

val instantiate_qformula_ty : t -> iqformula -> Typing.Ad_type.ad_uid Typing.Aast.aqformula

Instantiate a formula with a single type per sub-formula.

val check_type_var : iqformula -> unit

Raises `Wrong_modelling` if a variable in `TExists` has a type `CannotType`.