Module Typing__Tast

Typed AST (TAST). This is the AST used by the abstract domain because each sub-formula contains type information. In our context, the type of a formula is an index of an abstract element. We also have types for variables (see `Types.var_ty`) but they work at the value level rather than the level of abstract domains.

type tvariable = {

name : Lang.Ast.vname;
Name of the variable as appearing in the model.

ty : Core.Types.var_ty;
Type of the variable in the abstract element. If the element is a product, and the variable has two different abstract types in the sub-domains, then the concrete domain is stored.

uid : Typing.Ad_type.ad_uid;
UID of the abstract element in which the variable occurs.

}

val string_of_tvar : tvariable -> string

type tformula = Typing.Ad_type.ad_uid * tformula_

and tformula_ =

| TFVar of Lang.Ast.vname

| TCmp of Lang.Ast.bconstraint

| TEquiv of tformula * tformula

| TImply of tformula * tformula

| TAnd of tformula * tformula

| TOr of tformula * tformula

| TNot of tformula

type tqformula =

| TQFFormula of tformula

| TExists of tvariable * tqformula

Each variable has a type and belong to an abstract element.

val tformula_to_formula : tformula -> Lang.Ast.formula

val tqformula_to_qformula : tqformula -> Lang.Ast.qformula

val string_of_tformula : Typing.Ad_type.ad_ty -> tformula -> string

val string_of_tqformula : Typing.Ad_type.ad_ty -> tqformula -> string

val ctrue : tformula_

Encoding of true and false as formula.

val cfalse : tformula_

val ctrue' : tformula

val cfalse' : tformula

val ttrue : tqformula

val tfalse : tqformula

val vars_of_tformula : tformula -> Lang.Ast.vname list

val quantify : tvariable list -> tformula -> tqformula

quantify env f Given a variable environment `env`, existentially quantify the formula `f`. It adds `Exists` in front of `f` for each variable occuring in `f` and `env`. Variables not in `env` but in `f` (and vice-versa) are ignored.

val quantifiers : tqformula -> tvariable list

quantifiers tqf extracts all the existentially quantified variables from `tqf`. Duplicated quantifiers are removed. `Wrong_modelling` is raised if two identical variable's names have two distinct types.

val quantifier_free_of : tqformula -> tformula

quantifier_free_of tqf removes the quantifiers from the formula `tqf`.

val map_tformula : (tformula -> tformula) -> tqformula -> tqformula

map_formula f tqf applies the function `f` on `tqf` bypassing the quantifiers.

val merge_formula : (tformula -> tformula -> tformula) -> tqformula -> tqformula -> tqformula

merge_formula make f1 f2. Given two quantified formulas, merge their quantifier-free part using `make`. The existential binders of identical names are merged together.

val q_conjunction : Typing.Ad_type.ad_uid -> tqformula list -> tqformula

Merge a list of formula with the conjunctive connector of the given abstract domain type.

val q_disjunction : Typing.Ad_type.ad_uid -> tqformula list -> tqformula

val neg_formula : Typing.Ad_type.ad_uid -> tformula -> tformula

Negate a logic formula. NOT are pushed inwards such that it only occur over FVar. The size of the formula does not increase. The UID is the UID of a domain that support arbitrary logical formula.

val replace_uid : Typing.Ad_type.ad_uid -> tformula -> tformula

replace_uid uid tf replaces all the UIDs that are equal to `fst tf` by `uid` in `tf`.

val instantiate_vars : (Lang.Ast.vname * (tvariable * Bounds.Bound_rat.t)) list -> tformula -> tformula