Module Vardom.Bdd

TYPES

type t =

| T

| F

| N of t * t

You should never use the N constructor but `bdd_of` instead (to ensure maximal sharing)

type vardom_constraint = Typing.Tast.tvariable * Lang.Ast.cmpop * t

module B = Bounds.Bound_int

type bound = B.t

val name : string

val type_of : unit -> 'a

val bdd_of : t -> t -> t

val is_leaf : t -> bool

CONSTRUCTOR

val complete_bdd : int -> t

val zero : t

val one : t

val default_depth : int

val interpret : 'a -> 'b -> 'c

val to_formula : 'a -> 'b -> 'c

val top : ?⁠ty:Core.Types.var_ty -> unit -> t

val of_bounds : 'a -> 'b -> 'c

PRINTING and CONVERSIONS

val lb : t -> bound

val ub : t -> bound

val to_float_range : 'a -> 'b

val to_rational_range : 'a -> 'b

val to_range : t -> bound * bound

val depth_of : t -> int

val to_abstract_ty : t -> Core.Types.var_abstract_ty

val print : (string -> bound -> bound -> 'a, Stdlib.Format.formatter, unit) Stdlib.format -> t -> 'a

val to_expr : 'a -> 'b

SET-THEORETIC

val join : t -> t -> t

val meet : t -> t -> t Core.Bot.bot

val contains_float : t -> float -> bool

predicates

val is_singleton : t -> bool

val to_bot : t -> t Core.Bot.bot

val equal : 'a -> 'a -> bool

val float_size : 'a -> t -> float

mesure

val split : t -> t * t

Splits at the first possible depth

val split_on : t -> 'a -> t * t

val prune : 'a option

pruning TODO is it useful ?

val div : 'a -> 'b -> 'c

return valid values (possibly Bot, if dividend is nul)

val bdd_not : t -> t

val pref : t -> int -> t

val unop : Vardom_sig.unop_kind -> t -> t

val bdd_and : t -> t -> t

val bdd_or : t -> t -> t

val bdd_xor : t -> t -> t

val binop : Vardom_sig.binop_kind -> t -> t -> t

val eval_fun : 'a -> 'b -> 'c

function calls (sqrt, exp, ln ...) are handled here : given a function name and and a list of argument, it returns a possibly bottom result

FILTERING (TEST TRANSFER FUNCTIONS)

val filter_leq : 'a -> 'b -> 'c

val filter_lt : 'a -> 'b -> 'c

val filter_eq : t -> t -> (t * t) Core.Bot.bot

val filter_neq : 'a -> 'b -> 'c

val filter_div : 'a -> 'b -> 'c

val bot_if_false : t -> t Core.Bot.bot

val filter_unop : Vardom_sig.unop_kind -> t -> t -> t Core.Bot.bot

val filter_binop : Vardom_sig.binop_kind -> t -> t -> t -> t Core.Bot.bot * t Core.Bot.bot

val filter_fun : 'a -> 'b -> 'c

filtering function calls like (sqrt, exp, ln ...) is done here: given a function name, a list of argument, and a result, it remove points that cannot satisfy the relation : f(arg1,..,argn) = r; it returns a possibly bottom result

val filter_div_f : 'a -> 'b -> 'c -> 'd

val filter_root_f : 'a -> 'b -> 'c -> 'd

val filter_binop_f : Vardom_sig.binop_kind -> 'a -> t -> t -> t Core.Bot.bot