Module Trigo.Make

Parameters

I : Vardom_sig.S

Signature

include I

include Vardom_factory.Vardom_factory_sig

include Vardom_factory.Vardom_of_bounds

module B : Bounds.Bound_sig.S

type bound = B.t

type t

val of_bounds : ?⁠ty:Core.Types.var_ty -> (bound * bound) -> t * Core.Types.var_abstract_ty

Same as `Vardom_sig.top` but with bounds.

val of_bounds' : ?⁠ty:Core.Types.var_ty -> (bound * bound) -> t

val of_bound : bound -> t

val zero : t

val one : t

val minus_one : t

val zero_one : t

val minus_one_zero : t

val minus_one_one : t

val positive : t

val negative : t

val of_ints : int -> int -> t

val of_int : int -> t

val of_rats : Bounds.Bound_rat.t -> Bounds.Bound_rat.t -> t

val of_rat : Bounds.Bound_rat.t -> t

val of_floats : float -> float -> t

val of_float : float -> t

type vardom_constraint = Typing.Tast.tvariable * Lang.Ast.cmpop * t

See interpret.

val name : string

Name of the Vardom. See also `Abstract_domain.name`.

val type_of : unit -> Typing.Ad_type.vardom_ty

val interpret : Domains.Interpretation.approx_kind -> vardom_constraint -> t

Interpret the expression `v <op> constant` where constant has already been evaluated in this vardom. Returns the value of `v` fulfilling this constraint according to the approximation. Raise Wrong_modelling if the interpretation is not possible.

val to_formula : t -> Typing.Tast.tvariable -> Typing.Tast.tformula

Create a formula from the current vardom for the variable in parameter.

val top : ?⁠ty:Core.Types.var_ty -> unit -> t * Core.Types.var_abstract_ty

Top element (the less precise element) of the Vardom. In case of a concrete type, a suited abstract representation is picked, otherwise, the abstract type must be exactly the one supported in this vardom. Raise `Wrong_modelling` if the vardom cannot represent the given type.

val to_float_range : t -> float * float

Convert a vardom to a possibly over-approximated floating point range.

val to_rational_range : t -> Bounds.Bound_rat.t * Bounds.Bound_rat.t

Convert a vardom to a possibly over-approximated rational range.

val to_range : t -> bound * bound

Convert a vardom to a possibly over-approximated `bound` range.

val lb : t -> bound

Lower bound (smallest element in the vardom).

val ub : t -> bound

Upper bound (largest element in the vardom).

val print : Stdlib.Format.formatter -> t -> unit

val to_expr : t -> (Lang.Ast.cmpop * Lang.Ast.expr) * (Lang.Ast.cmpop * Lang.Ast.expr)

Convert the current values to its logical representation.

SET-THEORETIC

val join : t -> t -> t

val meet : t -> t -> t Core.Bot.bot

val contains_float : t -> float -> bool

predicates

val is_singleton : t -> bool

val to_bot : t -> t Core.Bot.bot

val equal : t -> t -> bool

val float_size : t -> float

mesure

val prune : (t -> t -> t list) option

pruning

INTERVAL ARITHMETICS (FORWARD EVALUATION)

val div : t -> t -> t Core.Bot.bot

return valid values (possibly Bot, if dividend is nul)

val unop : Vardom_sig.unop_kind -> t -> t

val binop : Vardom_sig.binop_kind -> t -> t -> t

val eval_fun : string -> t list -> t Core.Bot.bot

function calls (sqrt, exp, ln ...) are handled here : given a function name and and a list of argument, it returns a possibly bottom result

FILTERING (TEST TRANSFER FUNCTIONS)

val filter_leq : t -> t -> (t * t) Core.Bot.bot

val filter_lt : t -> t -> (t * t) Core.Bot.bot

val filter_eq : t -> t -> (t * t) Core.Bot.bot

val filter_neq : t -> t -> (t * t) Core.Bot.bot

val filter_div : t -> t -> t -> (t * t) Core.Bot.bot

val filter_unop : Vardom_sig.unop_kind -> t -> t -> t Core.Bot.bot

val filter_binop : Vardom_sig.binop_kind -> t -> t -> t -> (t * t) Core.Bot.bot

val filter_fun : string -> t list -> t -> t list Core.Bot.bot

filtering function calls like (sqrt, exp, ln ...) is done here : given a function name, a list of argument, and a result, it remove points that cannot satisfy the relation : f(arg1,..,argn) = r; it returns a possibly bottom result

val filter_div_f : t -> t -> t -> t Core.Bot.bot

Only filters the first argument

val filter_root_f : t -> t -> t -> t Core.Bot.bot

val filter_binop_f : Vardom_sig.binop_kind -> t -> t -> t -> t Core.Bot.bot

val pi_down : float

val pi_up : float

val sin_up : float -> float

val sin_down : float -> float

val cos_up : float -> float

val cos_down : float -> float

val tan_up : float -> float

val tan_down : float -> float

val atan_down : float -> float

val atan_up : float -> float

val acos_up : float -> float Core.Bot.bot

val acos_down : float -> float Core.Bot.bot

val asin_up : float -> float Core.Bot.bot

val asin_down : float -> float Core.Bot.bot

type float_itv = float * float

val print_fitv : Stdlib.Format.formatter -> (float * float) -> unit

val pihalf_down : F.t

val pihalf_up : F.t

val twopiup : float

val twopidown : float

val pihalf_fitv : F.t * F.t

val pi_fitv : F.t * F.t

val twopi_fitv : float * float

val fitv_to_i : (float * float) -> I.t

val pihalf_itv : I.t

val pi_itv : I.t

val twopi_itv : I.t

type monotony =

| Incr

| Decr

| Change of bool

val print_monotony : Stdlib.Format.formatter -> monotony -> unit

val itv : monotony -> ('a -> float) -> ('a -> float) -> ('a * 'a) -> float * float

val cosmonotony : (float * float) -> monotony

val cos_itv : I.t -> I.t

val acos_itv : I.t -> I.t Core.Bot.bot

val sin_itv : I.t -> I.t

val asin_itv : I.t -> I.t Core.Bot.bot

val tan_itv : I.t -> I.t Core.Bot.bot

val atan_itv : I.t -> I.t

val eval_fun : string -> I.t list -> I.t Core.Bot.bot

val normalize : float -> F.t -> I.t -> I.t * I.t

val arc : I.t -> (I.t -> I.t Core.Bot.bot) -> I.t -> I.t -> I.t Core.Bot.bot

val arcos_0_2pi : I.t -> I.t -> I.t Core.Bot.bot

val arcsin_mpih_pih : I.t -> I.t -> I.t Core.Bot.bot

val filter : I.t -> (I.t -> I.t -> I.t Core.Bot.bot) -> I.t -> I.t -> I.t Core.Bot.bot

val filter_cos : I.t -> I.t -> I.t Core.Bot.bot

val filter_sin : I.t -> I.t -> I.t Core.Bot.bot

val arctan_mpih_pih : I.t -> I.t -> I.t Core.Bot.bot

val filter_tan : I.t -> I.t -> I.t Core.Bot.bot

val filter_asin : I.t -> I.t -> I.t Core.Bot.bot

val filter_acos : I.t -> I.t -> I.t Core.Bot.bot

val filter_atan : I.t -> I.t -> I.t Core.Bot.bot

val filter_fun : string -> I.t list -> I.t -> I.t list Core.Bot.bot