Module Vardom__Vardom_factory

module type Vardom_of_bounds = sig ... end

module type Vardom_factory_sig = sig ... end

module Make : functor (V : Vardom_of_bounds) -> Vardom_factory_sig with type t = V.t and module B = V.B and type bound = V.bound

This module aims at extending a vardom with conveniency function derived from `of_bounds`. The variable type (`var_ty`) passed to `of_bounds` depends on the constructor used. For example, `of_ints` calls `of_bounds` with the type `Concrete Int`. A rule of thumb is that we always try to build the vardom with the concrete type of the arguments.