Module Vardom__Vardom_factory.Make

This module aims at extending a vardom with conveniency function derived from `of_bounds`. The variable type (`var_ty`) passed to `of_bounds` depends on the constructor used. For example, `of_ints` calls `of_bounds` with the type `Concrete Int`. A rule of thumb is that we always try to build the vardom with the concrete type of the arguments.

Parameters

V : Vardom_of_bounds

Signature

include Vardom_of_bounds

module B : Bounds.Bound_sig.S

type bound = B.t

type t

val of_bounds : ?⁠ty:Core.Types.var_ty -> (bound * bound) -> t * Core.Types.var_abstract_ty

Same as `Vardom_sig.top` but with bounds.

val of_bounds' : ?⁠ty:Core.Types.var_ty -> (bound * bound) -> t

val of_bound : bound -> t

val zero : t

val one : t

val minus_one : t

val zero_one : t

val minus_one_zero : t

val minus_one_one : t

val positive : t

val negative : t

val of_ints : int -> int -> t

val of_int : int -> t

val of_rats : Bounds.Bound_rat.t -> Bounds.Bound_rat.t -> t

val of_rat : Bounds.Bound_rat.t -> t

val of_floats : float -> float -> t

val of_float : float -> t