Module Vardom__Vardom_sig

Generic signature for variable domains, called `vardom` here. Examples of vardom are intervals of integers, floats, rational, sets or BDD. When implementing your vardom, consider using `Vardom_factory` to factorize some code.

type unop_kind =

| NEG

| ABS

| NOT

| PREF of int

| SUFF of int

Language of unary and binary operations that can be applied to `vardom`.

type binop_kind =

| ADD

| SUB

| MUL

| POW

| XOR

| AND

| OR

module type S = sig ... end

module type Vardom_functor = functor (B : Bounds.Bound_sig.S) -> S with module B = B